backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
scipy
requests
python-dotenv